make home directory compression more tunable, convert executables into subcommands, remove libclanker (#11)
* use longest home dir prefix possible

this is an issue on Arch Linux, for example, where the user bin has
home directory /. this causes trouble, so this change compresses based
on which user has the most specific/longest home directory.

* refactor compress::without_prefix

* only accumulate /etc/passwd entries whose home directories are
  prefixes of the given directory. this means we allocate less memory,
  and do less sorting work, at the cost of computing and storing the
  path without a prefix for potentially all entries of /etc/passwd
* remove mutable variables as much as possible. so buf and compressed,
  whose mutability was very underutilized, are replaced with variables
  on the spot. in retrospect, compressed is really the lossy utf-8
  representation of the compressed prefix, so this name is a little bit
  off.
* simplify logic where possible using PathBuf and String constructor
  functions and inline string literals

* refactor compress::without_prefix a second time

now compress::without_prefix doesn't sort the list of password entries.
it instead conducts a linear scan for the smallest element using the
same predicate function. this allows us to make
without_prefix_home_dir_and_username (holy long name, batman!)
immutable. we can use the same Vec::into_iter as well to avoid
reallocating our path.

* allow specifying minimum uid for home compression

this is typically 500 or 1000. i default to 0. in the future maybe this
will try and parse /etc/login.defs or something like that. i will come
to rue the day i said those words.

* remove unused use declarations from prompt.rs

* bump dependency versions, fix MaybeUninit UB

UB is fixed by only calling MaybeUninit::assume_init after the
associated libgit2 call succeeds. otherwise, we have uninitialized *mut
T floating around, which is undefined behavior in Rust. do we really
need all this unsafe, though? do i hate dependencies that much?

* remove libclanker

this was kind of a wishful thinking bit of cruft, wasn't it?

* impl trait color::Color for all T: Display

so it works for i32 as well as String or &str

* merge executables into subcommands, add argparsing

* `clanker-prompt` has become `clanker prompt`, with options
  -u,--unpriviliged-cursor, -p,--priviliged-cursor,
  -m,--min-home-dir-uid, and -w,--working-directory.
* `clanker-right-prompt` has become `clanker right-prompt`, with option
  -r,--return-code.
* `clanker-title` has become `clanker title`, with options
  -c,--currently-running, -m,--min-home-dir-uid, and
  -w,--working-directory.

note that for all of these, the current working directory can be
overridden from the command line, instead of only via the PWD
environment variable.

this makes deployment a little bit easier, at the cost of some argument
parsing overhead. honestly, probably not that big of a deal.

* change author email to me@gregjm.dev

this is my new preferred contact email

* remove unused `use std::env` from mod compress