rework home directory UID matching (#14)
* add max_home_dir_uid params to mod compress fns

this is handy because useradd(8), by default, makes human users in the
range 1000-60000. so on my system, user nobody has uid 65534.

add prompt argument -x,--max-home-dir-uid

specifies the maximum UID of a user who is elegible to be selected for
home-directory pathname compression

* increase minimum home directory UID to 1000

this lines up with the default value in /etc/login.defs. on macOS this
is 500, but i'm honestly only targeting myself with Linux.