v0.7.0 (#15)
* rework home directory UID matching (#14)

* add max_home_dir_uid params to mod compress fns

this is handy because useradd(8), by default, makes human users in the
range 1000-60000. so on my system, user nobody has uid 65534.

add prompt argument -x,--max-home-dir-uid

specifies the maximum UID of a user who is elegible to be selected for
home-directory pathname compression

* increase minimum home directory UID to 1000

this lines up with the default value in /etc/login.defs. on macOS this
is 500, but i'm honestly only targeting myself with Linux.

* add clanker-prompt arg -S,--no-username-hostname (#13)

if set, the current username and hostname of this machine will not be
printed by clanker-prompt. this is for esti, who for some reason has
requested this functionality

* remove trailing periods from argument help text

this matches the style used by clap in its documentation and help text
in autogenerated subcommands/arguments

* bump version to 0.7.0

breaking change when we modified the default value of
-m,--min-home-dir-uid. non-breaking change when we added the argument
-x,--max-home-dir-uid, and when we changed the help text

* s/priviliged/privileged/g

whoops, that's embarassing!